hwangsae: move common use functions to a separate file
