+++ hwangsae/relay.c
@@ -17,9 +17,8 @@
  */
 
 #include "relay.h"
+#include "common.h"
 
-#include <ifaddrs.h>
-#include <net/if.h>
 #include <srt/srt.h>
 #include <gio/gio.h>
 
@@ -533,52 +532,6 @@ hwangsae_relay_new (void)
 }
 
 static gchar *
-_get_local_ip (void)
-{
-  struct ifaddrs *addrs;
-  struct ifaddrs *it;
-  gchar *result = NULL;
-
-  if (getifaddrs (&addrs) < 0) {
-    return NULL;
-  }
-
-  for (it = addrs; !result && it; it = it->ifa_next) {
-    socklen_t addr_len;
-    char buf[INET6_ADDRSTRLEN + 1];
-
-    /* Ignore interfaces that are down, not running or don't have an IP. We also
-     * want to skip loopbacks. */
-    if ((it->ifa_flags & (IFF_UP | IFF_RUNNING | IFF_LOOPBACK)) !=
-        (IFF_UP | IFF_RUNNING) || it->ifa_addr == NULL) {
-      continue;
-    }
-
-    switch (it->ifa_addr->sa_family) {
-      case AF_INET:
-        addr_len = sizeof (struct sockaddr_in);
-        break;
-      case AF_INET6:
-        addr_len = sizeof (struct sockaddr_in6);
-        break;
-      default:
-        continue;
-    }
-
-    if (getnameinfo (it->ifa_addr, addr_len, buf, sizeof (buf), NULL, 0,
-            NI_NUMERICHOST) != 0) {
-      continue;
-    }
-
-    result = g_strdup (buf);
-  }
-
-  freeifaddrs (addrs);
-
-  return result;
-}
-
-static gchar *
 hwangsae_relay_make_uri (HwangsaeRelay * self, guint port)
 {
   g_autofree gchar *local_ip = NULL;
@@ -587,7 +540,7 @@ hwangsae_relay_make_uri (HwangsaeRelay * self, guint port)
   if (self->external_ip) {
     ip = self->external_ip;
   } else {
-    ip = local_ip = _get_local_ip ();
+    ip = local_ip = hwangsae_common_get_local_ip ();
   }
 
   return g_strdup_printf ("srt://%s:%d", ip, port);
